USed driver not reactive, now they are returning with a mono. added behaviours to our drivers.

+++ src/main/java/giuseppe/pinto/transportation/aggregator/domain/Trip.java
@@ -5,6 +5,7 @@ import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.ToString;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Currency;
 
@@ -20,7 +21,7 @@ public class Trip {
     private final LocalDateTime arrivalDate;
     private final String flightNumber;
     private final String airline;
-    private final Double price;
+    private final BigDecimal price;
     private final Currency currency;
     private final Driver driver;
 }
